ref(server): move command middleware to its own dir
